fixing 401 in routes

+++ routes/api.php
@@ -14,18 +14,18 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('api')->get('/user', function (Request $request) {
+Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
 
-Route::middleware('api')->get('books', 'App\Http\Controllers\BookController@index');
-Route::middleware('api')->post('books', 'App\Http\Controllers\BookController@store');
-Route::middleware('api')->get('books/{id}', 'App\Http\Controllers\BookController@show');
-Route::middleware('api')->put('books/{id}', 'App\Http\Controllers\BookController@update');
-Route::middleware('api')->delete('books/{id}', 'App\Http\Controllers\BookController@destroy');
+Route::middleware('auth:api')->get('books', 'App\Http\Controllers\BookController@index');
+Route::middleware('auth:api')->post('books', 'App\Http\Controllers\BookController@store');
+Route::middleware('auth:api')->get('books/{id}', 'App\Http\Controllers\BookController@show');
+Route::middleware('auth:api')->put('books/{id}', 'App\Http\Controllers\BookController@update');
+Route::middleware('auth:api')->delete('books/{id}', 'App\Http\Controllers\BookController@destroy');
 
 
 Route::post('login', 'App\Http\Controllers\AuthController@login');
 Route::post('register', 'App\Http\Controllers\AuthController@store');
-Route::middleware('api')->post('logout', 'App\Http\Controllers\AuthController@logout');
+Route::middleware('auth:api')->post('logout', 'App\Http\Controllers\AuthController@logout');
